# docs/swagger.yaml
basePath: /api/v1
definitions:
  request.CreateBanner:
    properties:
      content:
        description: Содержимое баннера
        type: object
      feature_id:
        description: Идентификатор фичи
        format: uint64
        type: integer
      is_active:
        description: Флаг активности баннера
        type: boolean
      tag_ids:
        description: Идентификаторы тегов
        items:
          type: integer
        type: array
    type: object
  request.UpdateBanner:
    properties:
      content:
        description: Содержимое баннера
        type: object
      feature_id:
        description: Идентификатор фичи
        format: uint64
        type: integer
      is_active:
        description: Флаг активности баннера
        type: boolean
      tag_ids:
        description: Идентификаторы тегов
        items:
          type: integer
        type: array
    type: object
  response.Banner:
    properties:
      banner_id:
        description: Идентификатор баннера
        format: uint64
        type: integer
      created_at:
        description: Дата создания баннера
        format: date-time
        type: string
      feature_id:
        description: Идентификатор фичи
        type: integer
      is_active:
        description: Флаг активности баннера
        format: uint64
        type: boolean
      tag_ids:
        description: Идентификаторы тэгов
        items:
          type: integer
        type: array
      updated_at:
        description: Дата обновления баннера
        format: date-time
        type: string
      versions:
        description: Последние три версии баннера
        items:
          $ref: '#/definitions/response.Content'
        type: array
    type: object
  response.BannerID:
    properties:
      banner_id:
        description: Идентификатор созданного баннера
        format: uint64
        type: integer
    type: object
  response.Content:
    properties:
      content:
        description: Содержимое баннера
        type: object
      created_at:
        description: Дата создания версии
        format: date-time
        type: string
      version:
        description: Версия содержимого баннера
        format: uint32
        type: integer
    type: object
  tools.Error:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: vet_v2002@mail.ru
    name: Артём Ветошкин
  description: Rest API Для управления для сервиса баннеров
  title: Сервис баннеров
  version: "1.0"
paths:
  /banner:
    get:
      description: Возвращает список баннеров на основе фильтра по фиче и/или тегу.
      parameters:
      - description: Идентификатор тэга группы пользователей
        in: query
        name: tag_id
        type: integer
      - description: Идентификатор фичи
        in: query
        name: feature_id
        type: integer
      - description: Лимит
        in: query
        name: limit
        type: integer
      - description: Оффсет
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список баннеров успешно отфильтрован
          schema:
            items:
              $ref: '#/definitions/response.Banner'
            type: array
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/tools.Error'
        "401":
          description: Пользователь не авторизован
        "403":
          description: Пользователь не имеет доступа
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - AdminToken: []
      summary: Получение всех баннеров c фильтрацией по фиче и/или тегу
      tags:
      - banner
    post:
      consumes:
      - application/json
      description: Добавляет баннер включая его содержания, id фичи, список id тэгов
        и состояние.
      parameters:
      - description: Информация о добавляемом пользователе
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateBanner'
      produces:
      - application/json
      responses:
        "201":
          description: Баннер успешно добавлен в систему
          schema:
            $ref: '#/definitions/response.BannerID'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/tools.Error'
        "401":
          description: Пользователь не авторизован
        "403":
          description: Пользователь не имеет доступа
        "409":
          description: Баннер с указанной парой id фичи и ia тэга уже существует
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - AdminToken: []
      summary: Создание нового баннера.
      tags:
      - banner
  /banner/{id}:
    delete:
      description: Удаляет информацию о банере по его id.
      parameters:
      - description: Идентификатор баннера
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Баннер успешно удалён
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/tools.Error'
        "401":
          description: Пользователь не авторизован
        "403":
          description: Пользователь не имеет доступа
        "404":
          description: Баннер с данным id не найден
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - AdminToken: []
      summary: Удаление банера.
      tags:
      - banner
    patch:
      consumes:
      - application/json
      description: Обновляет информацию о баннере по его id.
      parameters:
      - description: Идентификатор баннера
        in: path
        name: id
        required: true
        type: integer
      - description: Информация об обновлении
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateBanner'
      produces:
      - application/json
      responses:
        "200":
          description: Баннер успешно обновлён
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/tools.Error'
        "401":
          description: Пользователь не авторизован
        "403":
          description: Пользователь не имеет доступа
        "404":
          description: Баннер с данным id не найден
        "409":
          description: Баннер с указанной парой id фичи и ia тэга уже существует
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - AdminToken: []
      summary: Обновление баннера.
      tags:
      - banner
  /filter_banner:
    delete:
      description: Удаляет баннеры на основе фильтра по фиче или тегу. Обязателен
        один из query параметров.
      parameters:
      - description: Идентификатор тэга группы пользователей
        in: query
        name: tag_id
        type: integer
      - description: Идентификатор фичи
        in: query
        name: feature_id
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Баннеры успешно удалены
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/tools.Error'
        "401":
          description: Пользователь не авторизован
        "403":
          description: Пользователь не имеет доступа
        "404":
          description: Баннер с указанными тэгом и фичёй не найден
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - AdminToken: []
      summary: Удаление всех баннеров c фильтрацией по фиче или тегу
      tags:
      - banner
  /token/admin:
    get:
      description: Возвращает токен с правами админа.
      produces:
      - application/json
      responses:
        "200":
          description: Токен успешно создан
          schema:
            type: string
      summary: Получение токена админа.
      tags:
      - auth
  /token/user:
    get:
      description: Возвращает токен с правами пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: Токен успешно создан
          schema:
            type: string
      summary: Получение токена пользователя.
      tags:
      - auth
  /user_banner:
    get:
      description: '|'
      parameters:
      - description: Идентификатор тэга группы пользователей
        in: query
        name: tag_id
        required: true
        type: integer
      - description: Идентификатор фичи
        in: query
        name: feature_id
        required: true
        type: integer
      - description: Версия баннера
        in: query
        name: version
        type: integer
      - description: Получать актуальную информацию
        in: query
        name: use_last_revision
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: JSON-отображение баннера
          schema:
            type: object
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/tools.Error'
        "401":
          description: Пользователь не авторизован
        "403":
          description: Пользователь не имеет доступа
        "404":
          description: Баннер с указанными тэгом и фичёй не найден
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/tools.Error'
      security:
      - UserToken: []
      summary: Получение баннера для пользователя.
      tags:
      - banner
schemes:
- http
securityDefinitions:
  AdminToken:
    description: Токен доступа админа
    in: header
    name: token
    type: apiKey
  UserToken:
    description: Токен доступа пользователя
    in: header
    name: token
    type: apiKey
swagger: "2.0"